Add falling sand to main menu
Animated menu screens bring a lot of polish with relatively little
effort. In this case we want sand to be falling from the top of the
scene and piling up around the menu buttons.

Adds a new `MenuSand` class which populates a 2d array of Rectangles
covering the screen. We can change the opacity of these rectangles to
represent sand particles being in those positions.

Adds a Util namespace for some utility functions.

Adds a colour palette, and updates the menu buttons to use it.