Add music tracks
We wanted to have music that dynamically gets more intense as the
game progresses.

We've added 3 tracks which get played at the same time. The first
track is always on and is the background music, the second track gets
activated if the time remaining on any hourglass is less than 10
seconds, the third track gets activated when there's 5 seconds left.

Modifying the volume on a track is a light weight operation, so we
don't mind doing it to each track every frame, much better than
stopping and starting them, plus they stay in sync.